requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.35
python-dotenv==1.0.1
WTForms==3.1.2
Flask-WTF==1.2.1
pytest==8.3.2
